ccardimx28: separate DT into module and board
It is common practice to separate the DT in module part (dtsi)
and board part (dts).
This way, the ccardimx28.dtsi specifies the existing devices and
their pinctrl for this module. All devices that are not on the
module are disabled by default and need to be enabled in the
baseboard DT (ccardimx28js.dts).

https://jira.digi.com/browse/DEL-360

Signed-off-by: Hector Palacios <hector.palacios@digi.com>