ARM: mxs: ccardimx28 post_init takes ENET1 PHY out of reset
The ccardimx28 module has a pull-down on the PHYs reset lines which
keeps the PHYs reset until the GPIO line is pulled up. This prevents
the driver from detecting ENET1 PHY because, since ENET0 and ENET1
share the same MDIO bus, the function mdiobus_register() that makes
the PHY autodetection is only called once, upon ENET0 probe, and at
this point ENET1 PHY reset line is still pulled down.
This patch introduces a post_init function for the ccardimx28 platform
that pulls up ENET1 PHY reset line, so that its address is latched before
the driver tries to autodetect the PHYs.

https://jira.digi.com/browse/DEL-288

Signed-off-by: Hector Palacios <hector.palacios@digi.com>