ARM: dts: mxs: leave card detect out of common mmc pins config
MicroSD card sockets don't usually have card detect line. This pin
is actually not needed for the MMC to work and it is more of a
platform design decission to have it.
The card detect pin already has a configuration entry of its own:
'mmc0_cd_cfg' so we complete the iomux configuration here and let
platforms to include it or not depending on whether the card detect
line is routed to the SD socket.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit e3d4b7a3a8a850c965678603f70628d07572bc97)

Conflicts:
	arch/arm/boot/dts/imx28.dtsi