ARM: dts: ccardimx28: added GPIO SPI support
This patch adds gpio mux for using the three SSP ports exposed
in ccardimx28 module (SSP0, SSP1, SSP3) as GPIO SPI bit banging
ports.

By default, the ccardimx28js enables SSP3 and adds spidev slave
to it. SSP1 is disabled by default and SSP0 is left commented out
as the main function for this port is the uSD card.

https://jira.digi.com/browse/DEL-304

Signed-off-by: Hector Palacios <hector.palacios@digi.com>