ARM: dts: ccardimx28: add backlight support via PWM0
On JSK, PWM0 is routed to the LCD backlight.
This patch adds the IOMUX configuration for PWM0 and adds the
backlight device be driven by PWM0.

Since the backlight signal is inverted on the HR_LCD kit (0=on, 1=off)
and brightness-levels = <0 128 8 16 32 64 128 255>;
selecting an index 0 (0) will output a fully on backlight, while
an index of 6 (128) will put a 50% duty on the PWM. Notice that
this level may not be enough to light up the backlight.

Select by default an index 1 due to https://jira.digi.com/browse/DEL-484

This inverse login may be ammended by adding flags support to
the MXS PWM driver.

https://jira.digi.com/browse/DEL-296
https://jira.digi.com/browse/DEL-297
https://jira.digi.com/browse/DEL-480

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>