video: mxsfb: fix settings for 18bit data bus and 32bpp
For a selection of LCD data bus width of 18bits and a color
mode of 32bpp, the driver was setting the color mapping to
rgb666, which is wrong, as the color in memory realy has an
rgb888 layout.

This patch also removes the setting of flag CTRL_DF24 that
makes the driver dimiss the upper 2 bits when handling 32/24bpp
colors in a diplay with 18bit data bus width.
This made true color images display wrong in such configurations.

https://jira.digi.com/browse/DEL-482
https://jira.digi.com/browse/DEL-297

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Juergen Beisert <jbe@pengutronix.de>