ARM: mxs: add bluetooth support to ccardimx28
Add IOMUX Bluetooth control lines to module's Device Tree
attached to AUART0.

Request of GPIOs connected to the Atheros Bluetooth chip
is done in platform code because these don't belong to a
specific driver or Device Tree binding.
Besides, the power up of the Bluetooth chip must be done
after a delay.

https://jira.digi.com/browse/DEL-295

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>