ARM: mxs: register HWID into the Device Tree
The proper way to pass hardware information to the kernel
is through the Device Tree, so we'll encode the HWID as properties
of the main node. Properties will be prefixed with string "digi,hwid,"
followed by the name of each property:
 digi.hwid,tf
 digi,hwid,variant,
 digi,hwid,hv
 digi,hwid,cert
 digi,hwid,year
 digi,hwid,month
 digi,hwid,sn

By default, on a platform DTS these properties won't exist.
It will be responsibility of the boot loader to read the HWID from
the OTP bits (or from NVRAM) and embed these properties in the DT
before booting the kernel.

However, if the kernel cannot find these properties in the DT it will
retrieve the information from the OTP bits (not from NVRAM) and populate
them.
Then, if CONFIG_PROC_DEVICETREE=y in the kernel configuration,
these will be populated under /proc/device-tree/

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Javier Viguera <javier.viguera@digi.com>

https://jira.digi.com/browse/DEL-490