ARM: pm: imx28: Add machine specific suspend to RAM support.
Add platform code to allow for machine specific suspend to RAM.

Adding this to platform code is not ideal, but there is currently no
framework in the kernel to add power management support in drivers.

Also, the code that puts the EMI pads in low power mode is crashing
so it's commented out. It seems reading from the pinctrl virtual address
after the DRAM has been placed in self refresh is causing this. Before the
DRAM self refresh the pinctrl can be accessed fine, for example to light an
LED.

https://jira.digi.com/browse/DEL-323

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>