ARM: pm: imx28: Use stack to retrieve argument 1 in suspend to RAM.
Argument 1 passes a pointer to a structure keeping the virtual addresses
to the assembler suspend to RAM function.

The target will put the DRAM into self refresh, so the first thing the
function did was copy these addresses into registers.

This does not seem to be needed, so this change retrieves the pointer
to the virtual address from the stack.

https://jira.digi.com/browse/DEL-323

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>