ARM: mxs: pm: Move sram allocations to init function.
Given that gen_pool_free() hits a nasty BUG() due to arm not having an
NMI safe compare and swap implementation, move the allocation of the iram
pool to the initialization function and reuse it without freeing it.

Not sure why the BUG() is hit on a single core system if compare and swap
is typically used in synchronization on multicore systems, but the topic
is complex enough to stay away from it.

https://jira.digi.com/browse/DEL-613

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>