ARM: dts: ccardimx28: decrease the period of the PWM backlight
The PWM signal for the backlight was using a period of 5ms. This period
was too big for properly lighting the display with duty cycles below 50%
as the signal remained low for too much time.
Reducing the period to 15usec guarantees that the image is still visible
with duty cycles up to 1%.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Mike Engel <mike.engel@digi.com>

https://jira.digi.com/browse/DEL-628