ARM: dts: imx28-evk: Add pinctrl for USB OTG ID pin
After that, the dual-role switch can be used at imx28-evk.

Due to hardware limitations, it needs to switch VDD 5V
SOURCE SELECT to "USB 5V" if the user wants to use OTG
dual-role switch function. it needs to switch VDD 5V
SOURCE SELECT to "WALL 5V" if the user wants to Host 1.

Signed-off-by: Peter Chen <peter.chen@freescale.com>