ARM: cpufreq: ccardimx28js: Use clk framework internal clk detect.
The old driver used a machine specific clock framework call to detect
whether the USB and LCD clocks were enabled.

The common clock framework seems not to have an open API to check whether
a clk is used, and no current driver seems to use this.

We use __clk_is_enabled() which is accessible to drivers.

https://jira.digi.com/browse/DEL-324

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>