Revert "gpiolib: ccardimx28js: Add wakeup support."
Making gpiolib wakeup capable adds the power folder to the sysfs
gpio entries when exported. However, this folder is not meant to
be used in this way in this version of the kernel.

We had previously done this to keep the legacy enable/disable
functionality on specific gpios, but the work to do that now
is not as trivial.

Without this, gpio needs to be configured as IRQs in the standard way,
by setting an edge, and they can then be used to wake up the device.

https://jira.digi.com/browse/DEL-665

This reverts commit 29e2e8ff4ecbe4ecf28fab1df39bc4dd7ec52741.

Conflicts:

	drivers/gpio/gpiolib.c

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>