ARM: clk: divider: Round down current rate.
While testing the CPU frequency driver, trying to set the cpu frequency
was not working.

Calling clk_set_rate for 261818000 Hz changed the rate to 240MHz. The same
for 454MHz.

The problem was that reading the current rate returned 261818181, and the
now <= rate test failed. Rounding down to 261818000 solves the problem.

This hacked workaround needs to be investigated until a root cause is
found. Probably the problem is not with the clk framework but with
the cpufreq driver that needs to provide the profiles with higher
resolution, even using the working point profiles instead of a static
table as it was on the 2.6.35 driver.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>