ARM: cpufreq: Workaround a crash when changing frequencies.
All the cpu, emi and ahb clocks source from pll0, and all change the pll0
fractional divider to different values when changed.

We need to find out how the 2.6.35 driver behaved, and see if the profiles
table made pll0 use the same fractional divider in that case.

We might need to recalculate the profile table to use with the new clk
framework.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>