ENGR00279087-1 camera: enable mclk before read the camera ID
When the camera driver is built as module and done 'insmod' command,
the camera will not be detected. The error message is as follows.

$ insmod ov5640_camera.ko
ov5640 2-003c: cannot get io voltage
ov5640_read_reg:write reg error:reg=300a
camera ov5640 is not found

The reason is the mclk need to be enabled before read camera registers.
This patch fixes it.

To balance the usecount for the mclk, we need disable the mclk afterwards.

Signed-off-by: Robby Cai <R63905@freescale.com>