serial: mxs-auart: wait for FIFO to flush before shutdown
The shutdown function was not waiting for the FIFO (which may be the
real 16 byte FIFO or the DMA buffer, if DMA is enabled) to flush
before disabling the AUART.
This lead to many bytes not being transferred (specially at low
baudrates), as they were still in the DMA buffer when the AUART was
shutdown.
This patch also adds the check for the BUSY flag on the tx_empty
callback.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>