net: phy: workaround autonegotiation errata on Micrel KSZ80x1
Micrel PHY models KSZ8051, KSZ8031 and KSZ8021 suffer an errata by which
autonegotiation may occasionally fail to get the link after a reset of
the PHY (power on, hardware, software) or after restarting the
autonegotiation on the PHY.
The suggested workaround is to disable AN, wait 500usec, and then enable
again AN.

We implement this workaround in the generic PHY statemachine in the
PHY_NOLINK state, if the PHY id matches one of the affected models and
the AN was enabled. The workaround will apply only after the timeout
to wait for autonegotiation has expired (10 seconds).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Robert Hodaszi <robert.hodaszi@digi.com>

https://jira.digi.com/browse/DEL-620
(cherry picked from commit 9b2c86d268ed9da02e7446f2db907250e19841e2)