irqchip: mxs: add irq_set_wake support
Any enabled interrupt of the MXS is able to wake up the CPU.
The support to configure an interrupt as wakeup source is done
in software by maintaining a list with the wakeup interrupts enabled
from user space.

Upon the suspend call, a function is called that saves the enabled
interrupts and disables those that are not set as wakeup sources.
Upon resume, the original status of interrupts is restored.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>

https://jira.digi.com/browse/DEL-664
https://jira.digi.com/browse/DEL-667
https://jira.digi.com/browse/DEL-668
(cherry picked from commit 3e0a5cbe658f40d1e9868554123e4ce583291b11)