ARM: mxs: clear ALARM_WAKE flag before power-off
The ALARM_WAKE flag is set in persistent register 0 of the RTC if the
device was woken up by an RTC alarm. This flag persists until manually
cleared by the software, and the reason is that the software can take
any action it needs if the device was woken up by the RTC.
If it is not cleared and the target is powered-off it will restart the
target immediately. Since this is not wanted, clear it before powering
off.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>

https://jira.digi.com/browse/DEL-813
(cherry picked from commit a39e2d58b8a393ca9f94d567f434c01bc9bf4cd0)