serial: mxs-auart: ack fast CTS pulses on ISR
Fast pulses on the CTS line may lead to a situation where the
CTS is asserted and generates an interrupt, but when the ISR
is called, the CTS line has already been deasserted (this can
happen if the receiver is very fast processing data). In this
case, we must do nothing (the transmission must not stop).
The ISR must only consider the case where CTS is still asserted
while processing the interrupt it generated.

We have observed 100ns pulses when the receiver is a fast CPU (i.MX53)
at an intermediate baudrate of 230400.

This patch also removes some lines of the ISR that served to no
purpose (they were changing the istat local variable).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>

https://jira.digi.com/browse/DEL-848
(cherry picked from commit 77e4e84ab2847d2f27ac009a5c93676e979217c7)