serial: mxs-auart: enable DMA mode regardless of flow control
The driver was enabling DMA mode on the set_termios hook only when
hardware flow control was enabled. This was causing problems when
the flow control was changed from hard to none/soft because the DMA
was not correctly unconfigured.

This patch moves the DMA initialization to the startup hook to use
this mode regardless of the selected flow control.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>

https://jira.digi.com/browse/DEL-894
(cherry picked from commit 95fcb5a7b182e1f382131514161cc9ec33e96f17)