pwm-backlight: don't disable PWM for brightness==0
The driver was assuming that a brightness of 0 equates to disabling
the PWM. The PWM however may have inverse polarity and a brightness
of 0 may need to map to a steady high level output. Disabling the PWM
may put an unexpected voltage on the PWM pin instead.

This patch fixes this by disabling the PWM only when the lcd has been
blanked.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>

https://jira.digi.com/browse/DEL-484
(cherry picked from commit 33249ba4b3503d5a953c7f664b0018327d907d8d)