ARM: dts: ccimx6adpt: Configure HDMI iomux.
No extra configuration is needed for the HDMI interface on the ccimx6adpt.
Remove pinctrl_hdmi_hdcp_1 which was used on the sabre SD to configure
the I2C lines to the HDMI adapter. On the ccimx6adpt the HDMI is on I2C3
and the lines are already correctly configured.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>