driver: char: Add Freescale OTP driver.
This allows to read and write the efuses on i.MX6 and i.MX53 platforms.

It is also modified so that the ccimx6adpt reads the MAC1 efuse and adds
the hwid information to the device tree (if not already present).

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit c72318ceb8a9eddbe80bd2a32dbe270871725e5c)