drivers: input: touchscreen: Fusion 7/10: Modify for single touch.
The DEY user space has no understanding of the multi touch protocol. tslib
has no multitouch support, and is meant for resistive and not capacitive
touch, and QT 4.7  only has multi touch support when running over X.

This commit introduces a configuration option to convert the events
reported by the Fusion touch driver into the single touch protocol so
that the current user space can at least use the Fusion touch in that way.

An X based user space should be able to use the multi touch capabilities
but at this time it's untested.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>