ARM: dts: ccimx6adpt: Configure LDO2 to be always on.
The LDO2 regulator supplies the wireless and the bluetooth. This is the
same atheros chip but use different drivers.

On the mid term we will add driver support to enable and disable the
regulator, but for the time being the DA9063 default bootstrapping has
it on and all the drivers have to do is leave it as is.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>