gpio: mxc: Add support for wakeup interrupts.
This are a similar commit to the one done for the i.mx28. It enables the
wakeup interrupt on suspend and disables it on resume if the device
can_wakeup flag is set.

The hardware interrupt belongs to the GPIO port and not to individual
GPIOs, so once enabled any GPIO in the port may wake up the system.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>