net: fec: add barrier to transmit path to ensure proper ordering
Ensure that the writes to the descriptor data is visible to the hardware
before the descriptor is handed over to the hardware.

Having discussed this with Will Deacon, we need a wmb() between writing
the descriptor data and handing the descriptor over to the hardware.
The corresponding rmb() is in the ethernet hardware.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
(cherry picked from commit a29dc07047c66ae31354b3953e136356b2f92cf5)