ARM: dts: ccimx6adpt: Place usdhc4 as first device to be probed.
For the ccimx6adpt, usdhc4 corresponds to the eMMC. Digi's u-boot expects
this to be mmcblk0 in the kernel as it hardcodes the rootfs command line
argument to it when asked to boot from emmc.

It's unfortunate that the device ordering is on a file common to all
imx6qdl platforms, but the probe ordering should not affect any of the
imx6 based platforms.

https://jira.digi.com/browse/DEL-951

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>