drivers: gpio: da9063: Correct offset for hardware IRQ.
The irq_find_mapping() function expects a harware irq number constrained
to a specific irq domain. In the case of the PMIC the domain is registered
in da9063-irq.c (see the da9063_irqs enum). The base for this is not the
irq_base for the PMIC controller, but the offset in the da9063_irqs enum.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit c6a8c426e4ec74c6cb7f67de40fc624f72b2fe60)