ARM: dts: ccimx6adpt: Configure regulators for suspend.
The bcore1, bcore2, bperi and ldo11 regulators need to be kept enabled on
suspend. Also, their suspend voltage needs to be the same as the normal
state voltage. Otherwise, parts of the CPU which are needed to awake the
CPU are not powered.

This may change if the module power logic is modified.

Fixes https://jira.digi.com/browse/DEL-992

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>