sound: soc: codecs: sgtl5000: Enable the codec clock before accessing.
The sgtl5000_fill_defaults() function is a workaround to not having a
dedicated reset signal. It configures the codec with default values.

To be able to access the codec the clock needs to be enabled beforehand
so we can't wait until the soc layer does the enabling for us as the
codec needs to register beforehand.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>