ARM: dts: ccimx6: Change v4l2 configuration to a more sensible one.
By default, have a parallel camera in IPU0, CSI0 and a MIPI camera
in IPU1 CSI1. This should allow to capture from both cameras at the
same time, which if both cameras are in the same IPU is not possible.

Also, swap the order of MIPI and parallel camera detection on the I2C
bus with the hopes that the MIPI will be probed first.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>