drivers: media: mxc: capture: ov5640_mipi: Add digi,alt-pwr-gpios.
This is an alternative gpio for multiple uses. In our case, we have two
camera boards, one with a pwr down pin and no reset pin, and another one
with neither reset or pwr down pin. They both have an ov564x sensor which
by default have an i2c slave address of 0x3c, so there is a clash.

The driver with the digi,alt-pwr-gpios uses it to power down the other
camera sensor while it reconfigures its own i2c slave address.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>