fsl_otp: change width of month and s/n in HWID
The month is coded in hex and only needs a nibble (4 bits) for its
possible values (0x1..0xc), so make it occupy bits 23..20 of MAC0 and
extend the width of the serial number field to bits 19..0.

This length of Month and S/N is the same we used for ccardimx28 platform.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>