ccimx6: rename DTS files to contain the CPU family
There will be two CPU models of the ccimx6 platform:
 - Quad/Dual
 - DualLite/Solo

The current existing DTS files are for the Quad/Dual.
Following the naming convention of other iMX6 platforms, this
patch appends the 'q' to the imx6 prefix.

It also moves the entries in the Makefile to respect alphabetical
ordering of the list.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>