drivers: mfd: da9063_core: Fix power down mode.
This commit sets the PMIC to read its OTP values after power down. This is the
logical setup so that the PMIC is configured in the same way on a cold boot or
after a power up.

This also allows to wakeup with the power key after power off which was
not working.

In order for the power key wakeup after power down to work also after a
suspend, LDO11 and BPERI need to be configured not to follow the
sequencer.

https://jira.digi.com/browse/DEL-1052

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>