soc: codecs: sgtl5000: ccimx6: Disable power on codec on suspend.
The ccimx6sbc and ccimx6adpt are powering the sgtl5000 codec from the
bperi regulator which is shared with other peripherals. Hence, we cannot
disable the analog power on suspend on these platforms. What we can do is
user the power register to disable VAG, headphone, line out, DAC and ADC.

This and the previous soc commits fix https://jira.digi.com/browse/DEL-1082

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>