ARM: dts: ccimx6: Set default bpp for HDMI to 32.
This is to avoid the X server having to set the framebuffer to 32 which
hits a bug on console switch (or suspend/resume) that results on the
framebuffer image being distorted.

See https://jira.digi.com/browse/DEL-1120

This commit was not originally applied to v3.10/master as the problem
only happened with HDMI as the primary.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit 8b334e6d303766ce7df975ad5858699fec649a03)

Conflicts:
	arch/arm/boot/dts/imx6q-ccimx6adpt.dts