serial: imx: disable automatic hw-flow when throttle requested by tty
If the tty buffers are close to full, the tty will call throttle() which
eventually calls the driver's set_mctrl() hook to assert RTS (CTS in
Freescale nomenclature) to tell the transmitter to stop sending data.
When automatic RTS control is enabled (CTSC) the driver cannot assert RTS
and the tty buffers may overflow.
To avoid this situation, we must momentarily disable the automatic hw-flow
control (clearing CTSC) and assert RTS (set CTS) until the buffers are
read out by the user space and the tty calls unthrottle() in which case
we can return to automatic hw-flow control (by enabling CTSC).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-973