serial: imx: add 2-wire RS-485 half duplex support
This patch adds 2-wire RS-485 half duplex support. The CTS line (output
on the i.MX6) is used as standard RTS to control the direction of the
data (to enable/disable the RS-485 drivers).
Instead of using the CTS line as a GPIO, the UART provides a way to control
the status of the line manually, so the UART ports IOMUX must be configured
as three lines: RX/TX/CTS.
On stop_tx(), we must wait for the data to be shifted out from the shift
register before deactivating RTS. For that, we need to poll the USR2_TXDC
flag. With the previously existing delay of 5usec this flag was being set
before the data was fully shifted out (at least at baudarates <= 9600) so
we calculated a udelay for the time it takes to transmitt one bit at the
programmed baudrate, to guarantee the transmission was really complete.
As security measure, a timeout is used to avoid a kernel hang in this loop.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-753
Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>