Revert "drivers: video: mxc_ipuv3_fb: Show logo on all framebuffers."
This reverts commit 8134bce3dd29f862bc693d82a025727083b49990.

This was meant to be a single change that would work with all display types.
Unfortunately it did not work for the HDMI driver which has a late init due
to the hotplugging support.

So we had patched the HDMI driver to setup on framebuffer registration. After
extensive reboot testing, that change has undesired effects on system reboot,
so it is also being reverted.