imx: imx6q: ccimx6: Add a board version getter.
The ccimx6 module uses an OTP to store a carrier board version number. This
is either read by the u-boot bootloader or the fsl_otp driver and
dynamically inserted into the device tree.

It is used to distinguish between hardware board revisions (not module
revisions), when the differences are so small a different device tree
file is overkill.

This commit will add a mx6q_get_board_version() that will read the board version once
from the device tree and store it for future use.

Use this function from drivers needing to do conditional actions depending
on the board version. It can also be used by carrier boards based on the
ccimx6 reference designs to introduce small variations without the need
to create a new device tree file.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>