ARM: ccimx6sbc: fixup DT dynamically for audio power on SBCv2
SBCv2 controls the power enable of the audio codec using GPIO2_25.
This patch asumes SBCv2 is the default version of the hardware and
adds a fixed regulator controlled by GPIO2_25.
On the init platform code, the patch queries the carrier board version
number and if it is an SBCv1, reconfigures the codec supplies in the DT
to source from 'bperi' regulator instead (always on).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>

https://jira.digi.com/browse/DEL-1233