drivers: video: mxc: Fix race condition on hdmi hotplug.
The intention was to display logos on all displays with the minimal changes
required, which has usually been easy.

I do not like that the mxc_hdmi driver is not setup when fbmem requests
for the logo to be shown, nor that we could not do this from
the mxc_ipuv3_fb code without introducing reboot issues. And even less that
there are race conditions between the display driver and user space when
doing it from the hdmi driver.

So this is about avoiding the race condition to show the logo without
having to rewrite part of the hdmi driver, and it is not pretty.

https://jira.digi.com/browse/DEL-1283

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>