MLK-10051-2 usb: common: otg: set feature of b_hnp_enable after host request flag is set
The A-device is required to set this feature and suspend the bus within
THOST_REQ_SUSP when it determines that the B-device wishes to become host
(host_req_flag = TRUE). So this patch does this if host request flag is set
and a_set_b_hnp_en has not been set.

Acked-by: Peter Chen <peter.chen@freescale.com>
Signed-off-by: Li Jun <jun.li@freescale.com>