MLK-10089 arm: imx: Add 198MHz OPP for i.MX6SX
Update the i.MX6SX operating points to comply with the latest
datasheet. Latest i.MX6SX datasheet of Rev.F, 1/2015 adds the
198MHz setpoint. For the RevB board, the VDD_ARM and ADD_SOC
are connected together, so the voltage for 198MHz needs to be
set to 1.175V. for the general setting, add a 25mV margin to
cover the board IR drop.

this patch is copy from commit: 17dae6e92 on branch imx_3.10.y
as too many conflicts to resolve.

Signed-off-by: Bai Ping <b51503@freescale.com>