ARM: dts: ccimx6sbc: Review CPU and SoC voltages for bypass.
Review the CPU and SoC voltages according to Dialog's reference report
(AN-PM-053).

Note that the SoC voltages do not vary. According to Dialog this is a
hardware limitation. By not changing it we avoid a SW crash (Dialog does
not give specific details on the crash).

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit 685e80cd0a9657c1dcb28c604be8eb24de533137)