drivers: media: mxc: ov5640_mipi: Update I2C address on software reset.
When the I2C configured in the DT is different from the default, the
driver configures itself with it via the I2C bus. However, when
performing a software reset the address goes back to the default value.

This commits updates the I2C address to the value in the DT after the
software reset command is issued.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>