MLK-10403 mmc: sdhci-esdhc-imx: remove invalid HS400 warning message
Currently, for i.MX7D-SDB/i.MX7D-12x12-arm2 board, when eMMC change
to HS400 mode, it will print following warning message:

sdhci-esdhc-imx 30b60000.usdhc: warning! HS400 strobe DLL status REF not lock!
sdhci-esdhc-imx 30b60000.usdhc: warning! HS400 strobe DLL status SLV not lock!

This warning is print by the function esdhc_set_strobe_dll(), if the
eMMC support HS400, this function should be called.

But due to the HS400 mode initialization need a special operation as
following:
1)   go into HS200 mode;
2)   convert from HS200 mode to HS mode:  set clock to 52MHz,
	 and go into High speed mode;
3)   chose to be HS400 mode, and set clock to 200MHz.

All these set clock operation will finally call the function
esdhc_set_strobe_dll(),  but the first time this function be called,
the card clock is 52MHz, in this clock rate, the strobe_dll_status
register can’t be locked. If the clock is too slow, the ½ cycle is
too long, this make the length of a delay cell is not enough.
The second time when the clock rate is much higher, the
strobe_dll_status register can be locked normally.

We don't care the warning message when the clock is 50MHz, we just
care that whether the strobe DLL status REF/SLV is locked when the
card clock is 200MHz, this is the real warning message.

Signed-off-by: Haibo Chen <haibo.chen@freescale.com>