ASoC: imx-sgtl5000: enable DAC only when headphone is connected
On some hardware, plugging of the headphone jack can cause a short that
can make its way into the codec, causing it to fall back to a mode where
it does not generate further I2S clocks.
By disabling the DAC when the jack is not inserted, we can isolate the
codec from suffering these shorts during insertion of the jack.

Still the issue may appear during disconnection of the jack (while the DAC
is enabled). This can only be corrected by adding filters by hardware.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-1572