ASoC: imx-sgtl5000: add debounce histeresys to headphone jack detection
On some hardware, a short may happen during the insertion/removal of the
headphone jack. This short may make its way into the codec and turn it
unstable. Disabling the DAC when the jack is removed prevents the issue
from happening during insertion (the DAC is enabled after debounce time).
This commit adds some histeresys to the process by setting to zero the
debounce time after jack insertion, which will disable the DAC as fast as
possible during removing of the jack.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-1572