MLK-10660-3 video: mxc ipuv3 fb: Correct prefetch flag if no PRE is available
For those SoCs which have no PRE engines embedded, users may try to call
the ioctrl MXCFB_SET_PREFETCH to enable the PRE.  The ioctrl sets the flag
mxcfbi->prefetch to be true to go for PRE allocation.  It is doomed to fail
and the users will have no chance to set the flag back to false.  As a result,
the framebuffer will not be responsive to the users' set_par ioctls.  Let's
set the current prefetch flag back to the prefetch flag in that case to avoid
the situation.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>