MLK-10782-1 ARM:imx6qdl:Add support to enable/disable MMDC_CH1 handshake
MMDC CH1 is not used on i.MX6Q, so the handshake needed to change the parent
of periph2_sel or the divider of mmdc_ch1_axi_podf will never succeed.
Disable the handshake mechanism to allow changing the frequency of mmdc_ch1_axi,
allowing to use it as a possible source for the LDB DI clock.

Signed-off-by: Philipp Zabel <p.zabel at pengutronix.de>
Signed-off-by: Ranjani Vaidyanathan <Ranjani.Vaidyanathan@freescale.com>