MLK-10804-01 ARM: imx: update the ldb_dix clk info in clock tree on i.mx6ul
According to the design team's feedback, the 'CG6' in CCGR3 is for gpio_4's
clk, no 'CG' is for 'LDB_DI0' clk, so remove this gate2 clk from the clock
tree. The 'ldb_di0' and 'ldb_di1' input for lcdif_clk_sel and sim_clk_sel is
the clock derived from the 'ldb_di0_div_sel' and 'ldb_di1_div_sel', so rename
these clock to 'ldb_di0' and 'ldb_di1' to align with the name used in the DOC.
Additionally, for 'ldb_di1_div_7' and 'ldb_di1_div_3_5', the parent clock is
qspi1_sel.

Signed-off-by: Bai Ping <b51503@freescale.com>