MLK-10837 dts: imx6qp-sabreauto: enable cut power of usdhc3 slot
On old i.MX6Q/DL Sabreauto board, we can not use external regulator for
card slot by default since the card power is shared with card detect pullup.
Disabling the vmmc regulator will cause unexpected card detect
interrupts. HW rework is needed to fix this isssue.
So the using of external regulator is disabled by default in dts as follows:
/* vmmc-supply = <&reg_sd3_vmmc>; */

On imx6qp-sabreauto board, this HW issue is fixed already, so enable it
in imx6qp-sabreauto dts to use this function.
This is needed for a SD3.0 card re-enumeration working on UHS mode.
Otherwise the card may be re-enumerated as SD2.0 mode after system
suspend/resume due to not cut off power/reset card.

Signed-off-by: Dong Aisheng <b29396@freescale.com>