ARM: dts: ccimx6sbc: Add iomux configuration to audio power toggle GPIO.
On the SBCv2 we can control the 3.3V supply of the audio codec through
GPIO2_25. The pin default IOMUX configuration is not GPIO, so it needs
to be correctly configured in order to correctly toggle the line.

The sgtl5000 audio codec in particular has been seen to sometimes leave
the I2C data line in a busy state on reboot if the power line is not
toggled.

Related issues:
https://jira.digi.com/browse/DEL-1224
https://jira.digi.com/browse/DEL-1570
https://jira.digi.com/browse/DEL-1572
https://jira.digi.com/browse/ADK4A-153

https://jira.digi.com/browse/ADK4A-153
(cherry picked from commit a19324775ac57e145672f7fc1a4492674c262de9)

Signed-off-by: Mike Engel <mike.engel@digi.com>
Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>