ARM: dts: Empty memory node to avoid confusion.
The contents of the memory node do not represent the actual memory
configuration of each variant, it's the bootloader that sets it on the
fly.

This commit removes the node contents and leave it just as a placeholder to avoid
confusion.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit f8f910da6ec6e109e79192623136d1fc6c7d3112)