MLK-10922 dts: imx6ul: uSDHC should be derived from mx6sx
MX6UL uSDHC performance is bad which is caused by ADMA is not enabled.
That is because the uSDHC is set to derive from MX6SL in device tree
and MX6SL ADMA is disabled by default due to an ADMA IC bug.
Actually the mx6ul uSDHC is derived from mx6sx, so fixing it and
enable ADMA.

Signed-off-by: Dong Aisheng <b29396@freescale.com>