MLK-10862-2 mxc IPUv3: cpmem: SMFC chan 12bit 3 planar burst size fixup
There are some little horizontal stripes on the I420 interlaced frames
captured by i.MX6QP SMFC channels but not i.MX6Q Rev1.  Changing the
IDMAC burst size from 32pixel to 16pixel may workaround this issue which
is recorded by TKT266260.  The capture performance is expected to be
dropped as the burst size is half.  The exact performance drop data is
not known since the video input device's output is not tough enough to
push the capture to it's boundary.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>