MLK-11066 video: mxc ipuv3 fb: check tile block after blank check in pan display
We may wrongly check tile block in pan display for a linear framebuffer in the
following case:
step 1) framebuffer works in linear RGB24 pixel format(unblanked)
step 2) blank the framebuffer
step 3) change framebuffer to be a GPU tile format
step 4) change framebuffer back to linear RGB24 pixel format

The cached variable mxc_fbi->cur_var will not change in all of the 4 steps
because the framebuffer is blanked in step 2).  In step 3), the format change
is successful since step 2) to 3) would definitely changes the pixel format
from linear RGB24 to a GPU tile format - mxc_fbi->cur_var is not the same
to fbi->var, which leads to mxcfb_need_to_set_par() returns true.  However,
in step 4), the format change is not successful since mxc_fbi->cur_var and
fbi->var is the same, which leads to mxcfb_need_to_set_par() returns false.
Note that we set mxc_fbi->resolve to be true in step 3), so it keeps to be
true in step 4) as we return at mxcfb_need_to_set_par() before we may change
it to be false.  Finally, in step 4), after doing ->fb_set_par() in
fb_set_var(), we do fb_pan_display() and wrongly check tile block for the
linear RGB24 pixel format as mxc_fbi->resolve is true - hit the BUG_ON() in
fmt_to_tile_block().  To fix this issue, this patch moves the code for tile
block check down below blank check in ->fb_pan_display().

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>