bluetooth: hci_ath: Stop filtering TX SCO packages.
The driver was initially intended for the AR3001 chip that does not
support the SCO HCI link.

Newer chips like the AR3002 support it and need it for HCI to work.

This was originally submitted to the bluez kernel development list
but never made it upstream.

http://thread.gmane.org/gmane.linux.bluez.kernel/10444

The original patch had a continuation which prioritized SCO packages
which did not work on our testing and it was dropped and not looked into
any further.

https://jira.digi.com/browse/DEL-1144

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>