ARM: defconfig: ccimx6sbc: Add IMX6SX to multi-arch build.
This is to avoid link errors of the type:

arch/arm/mach-imx/built-in.o: In function `clk_pllv3_do_shared_clks':
:(.text+0x118c): undefined reference to `amp_power_mutex'
:(.text+0x1190): undefined reference to `amp_power_mutex'
:(.text+0x11a0): undefined reference to `shared_mem'
:(.text+0x11a4): undefined reference to `shared_mem'
:(.text+0x11b4): undefined reference to `imx_sema4_mutex_lock'
:(.text+0x11e0): undefined reference to `imx_sema4_mutex_unlock'
:(.text+0x1218): undefined reference to `imx_update_shared_mem'

The FSL BSP is compiling common clk code with symbols which are only
present on the IMX6SX architecture.

Adding the IMX6SX is closer to the FSL BSP than fixing the errors for the
time being.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>